Allow `jlink` to generate a stripped down JRE without the `runtime` fileset
As a bonus, this fixes issue #30.  This works with AdoptOpenJDK- I didn't try
HomeBrew, because apparently the naming convention isn't supported (issue #36).
The `runtime` fileset doesn't work for stripping down jdk-11 because while the
individual modules can be included or excluded, they seem to be ignored when the
app is run, and it wants the monolithic 'modules' file instead.  That is ~134MB.

The project files are more readable and diffable with individual jmod lines than
a single comma delineated list, so that's what this implements.  It could be
smarter and run `jdeps` on each *.jar in the classpath, and hand that off to
`jlink`.  But this is simpler.  I don't see a need to run `jlink` to a temporary
directory and then pass it through the `runtime` filter, so they are mutually
exclusive.

The `runtime` fileset doesn't carry over the man pages, headers, or binaries
AFAICT, and I don't see the use for them in a Mac app, so they are turned off.

Example usage:

    <appbundler ...>
      <!-- runtime is the same value as `dir` in the `runtime` nest element -->
      <jlink runtime="/path/from/java_home">
        <jmod name="java.base"/>
        <jmod name="java.desktop"/>
        <jmod name="java.naming"/>
        <jmod name="java.perf"/>
        <jmod name="java.rmi"/>
        <jmod name="java.sql"/>
      </jlink>
      ...
    </appbundler>